+++ assets/scripts/pages/user/body/home/contact/index.tsx
@@ -7,32 +7,13 @@ import {
 } from "@material-ui/core/styles";
 
 
-
-import {Title} from "../../../common";
+import {Title, Container} from "../../../common";
 
 import ContactForm from "./contact-form";
 import ContactText from "./contact-text";
 
 const useStyles = makeStyles((theme:Theme) => {
 	return createStyles({
-		root: {
-			display: "flex",
-			flex: "0 0 3.125rem",
-			justifyContent : "center",
-			backgroundColor:(value)=> value?(theme.palette.type==="light")?"#F2F6F9":"#19191A":"",
-			margin:theme.spacing(4,0),
-			
-		},
-		container : {
-			display:"flex",
-			flex: "1 1 100%",
-			margin:theme.spacing(1,1),
-			padding:theme.spacing(3,0),
-			[theme.breakpoints.up("lg")]: {
-				display:"flex",
-				flex: `0 0 ${theme.breakpoints.values.lg}px`,
-			},
-		},
 		section : {
 			display: "flex",
 			flexDirection:"column",
@@ -53,21 +34,19 @@ const useStyles = makeStyles((theme:Theme) => {
 });
 
 const Contact = ({dark}) => {
-	const classes = useStyles(dark);
+	const classes = useStyles();
 	return (
-		<div className={classes.root}>
-			<div className={classes.container}>
-				<div className={classes.section}>
-					<div className={classes.title}>
-						<Title title={"Contact Us"} dark={false}/>
-					</div>
-					<div className={classes.content}>
-						<ContactText />
-						<ContactForm />
-					</div>
+		<Container dark={dark}>
+			<div className={classes.section}>
+				<div className={classes.title}>
+					<Title title={"Contact Us"} dark={false}/>
+				</div>
+				<div className={classes.content}>
+					<ContactText />
+					<ContactForm />
 				</div>
 			</div>
-		</div>
+		</Container>
 	);
 };
