+++ assets/scripts/pages/user/body/home/tutorials/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import {Title} from "../../../common";
+
 import {
 	Theme,
 	makeStyles,
@@ -7,6 +7,7 @@ import {
 } from "@material-ui/core/styles";
 
 
+import {Title, Container} from "../../../common";
 import LatestVideos from "./latest-videos";
 import PopularVideos from "./popular-videos";
 
@@ -14,39 +15,25 @@ const useStyles = makeStyles((theme:Theme) => {
 	return createStyles({
 		root: {
 			display: "flex",
-			flex: "0 0 3.125rem",
-			justifyContent : "center",
-			backgroundColor:(value)=> value?(theme.palette.type==="light")?"#F2F6F9":"#19191A":"",
-			margin:theme.spacing(2,0),
-		},
-		container : {
-			display:"flex",
-			flex: "1 1 100%",
-			margin: theme.spacing(1,0),
-			[theme.breakpoints.up("lg")]: {
-				display:"flex",
-				flex: `0 0 ${theme.breakpoints.values.lg}px`,
-			},
+			flexDirection:"column"
 		}
 	});
 });
 
 const Tutorials = ({dark}) => {
-	const classes = useStyles(dark);
+	const classes = useStyles();
 	return (
-		<div className={classes.root}>
-			<div className={classes.container}>
+		<Container dark={dark}>		
+			<div className={classes.root}>
 				<div>
-					<div>
-						<Title title={"Tutorials"} dark={false}/>
-					</div>
-					<div>
-						<LatestVideos />
-						<PopularVideos />
-					</div>
+					<Title title={"Tutorials"} dark={false}/>
+				</div>
+				<div>
+					<LatestVideos />
+					<PopularVideos />
 				</div>
 			</div>
-		</div>
+		</Container>
 	);
 };
