+++ assets/scripts/pages/user/common/index.tsx
@@ -1,7 +1,9 @@
 import Title from "./title";
 import FollowUs from "./follow-us";
+import Container from "./container";
 
 export {
+	Container,
 	FollowUs,
 	Title
 };
